% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{.read_enron_block}
\alias{.read_enron_block}
\title{Read and clean a block of Enron spreadsheet data}
\usage{
.read_enron_block(start_row, start_col, end_col, path, locations)
}
\arguments{
\item{start_row}{Integer scalar. The first row to read (the row with the
date).}

\item{start_col}{Character scalar. The first column to read.}

\item{end_col}{Character scalar. The last column to read.}

\item{path}{The path to the spreadsheet.}

\item{locations}{The character vector of locations.}
}
\value{
A clean tibble.
}
\description{
In a real package, I like to document even internal functions, to remind
future me how things work. Note the \verb{@keywords internal} bit at the bottom. I
also like to prefix unexported functions with \code{.}, again to help future me
remember that it's not supposed to be exported. I'm just documenting this
single internal function to give the idea.
}
\keyword{internal}
