% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{download_github}
\alias{download_github}
\title{Download enron data from GitHub}
\usage{
download_github(link, destfile)
}
\arguments{
\item{link}{Character. Link to the \code{.xls} enron file.}

\item{destfile}{Character. Filepath where the \code{.xls} file should be stored.}
}
\description{
Download enron data from GitHub
}
\details{
This is a helper function that gets called when\code{source = "github"}.
To download the file from GitHub, the string \code{?raw=true} needed to be added.
}
