% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enron}
\alias{enron}
\title{cleaned enron dataset}
\format{
rda with 4 columns

\describe{
\item{\code{location}}{character location}
\item{\code{date}}{character}
\item{\code{deliveries}}{integer}
\item{\code{receipts}}{integer}
}
}
\source{
Streamline
}
\usage{
enron
}
\description{
cleaned enron dataset
}
\keyword{datasets}
