% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_data}
\alias{enron_download_data}
\title{Download Enron spreadsheet data}
\usage{
enron_download_data(path = "./enron.xls", verbose = FALSE)
}
\arguments{
\item{path}{Where to download the spreadsheet. Defaults to a file named
"enron.xls" in the working directory.}

\item{verbose}{If \code{TRUE}, messages will be relayed from \code{\link[=download.file]{download.file()}}.}
}
\value{
The path, invisibly, for use in pipelines.
}
\description{
Download the Henry Hub tracking sheet, an extremely untidy spreadsheet that
was extracted from the \href{https://en.wikipedia.org/wiki/Enron_Corpus}{Enron Emails Corpus}. More information on this
data is available from \href{https://www.felienne.com/archives/3634}{Felienne Hermans}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- tempfile(fileext = ".xls")
enron_download_data(path)
unlink(path)
\dontshow{\}) # examplesIf}
}
