% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_file}
\alias{enron_download_file}
\title{Downloads the Enron dataset from the 'StreamlineDataScience/enron-example'
GitHub repository.}
\usage{
enron_download_file(path)
}
\arguments{
\item{path}{The destination file path for the downloaded data. If missing, a temporary location is used.}
}
\value{
Returns the file path to the downloaded dataset (invisibly).
}
\description{
Downloads the Enron dataset from the 'StreamlineDataScience/enron-example'
GitHub repository.
}
\examples{
enron_download_file()
}
