% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_sample_data.R
\docType{data}
\name{enron_sample}
\alias{enron_sample}
\title{Sample Enron Data for Testing and Analysis}
\format{
A \code{tibble} with 4 columns, which are:
\describe{
\item{location}{The location that receipts and deliveries are for }
\item{date}{The date that receipts and deliveries are for }
\item{deliveries}{Delivery totals }
\item{receipts}{Receipt totals }
}
}
\usage{
enron_sample
}
\description{
A \code{tibble} that is for testing and running examples
}
\keyword{datasets}
