% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_tidy_data.R
\docType{data}
\name{enron_tidy_data}
\alias{enron_tidy_data}
\title{Enron deliveries and receipts data}
\format{
\subsection{\code{enron_tidy_data}}{

A data frame with 2,054 rows and 4 columns:
\describe{
\item{location}{Location}
\item{date}{Date in YYYY-MM-DD format}
\item{deliveries}{deliveries}
\item{receipts}{receipts}
}
}
}
\source{
\url{https://github.com/StreamlineDataScience/enron-example/blob/main/andrea_ring_000_1_1.pst.0.xls}
}
\usage{
enron_tidy_data
}
\description{
Tidied data from and Enron XLS spreadsheet
}
\keyword{datasets}
