% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(path, type = c("receipts", "deliveries"), location_names, ranges)
}
\arguments{
\item{path}{Character. Filepath of the \code{.xls} file to be processed.}

\item{type}{Character. The type of data being extracted. Possible values are
\code{"receipts"} and \code{"deliveries"}.}

\item{location_names}{Tibble. A one-column tibble where each row corresponds
to a location name. The name of the column is \code{location}.}

\item{ranges}{List. A list of two elements: \code{date_row} and \code{data_range}. Each
of these elements expect the cells where the corresponding data is stored.}
}
\description{
Get data
}
\details{
This helper function is used to process receipts and deliveries data.
It expects each set of cell ranges to be expressed in its own element of a
list. Dates are used as column names for each set of data. These dates need to
be processed in order to be properly used. The location names are binded to
wide data, which is later pivoted to long format.
}
