% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{get_default_link}
\alias{get_default_link}
\title{Get default download link}
\usage{
get_default_link(source)
}
\arguments{
\item{source}{Character. Indicates where to download enron data from.
Possible options are \code{"github"} and \code{"google_sheets"}.}
}
\description{
Get default download link
}
\details{
This function is used to get download links when \code{link = NULL}.
Default \code{link}s for different \code{source}s were defined in \code{R/constants.R}.
}
