% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{get_location_names}
\alias{get_location_names}
\title{Get location names}
\usage{
get_location_names(path, range)
}
\arguments{
\item{path}{Character. Filepath of the \code{.xls} file to be processed.}

\item{range}{Character. Cells where location names can be found in the file.}
}
\description{
Get location names
}
\details{
This helper function is used to extract the set of location names.
}
