% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.rdev-package.R
\docType{package}
\name{str.rdev-package}
\alias{str.rdev}
\alias{str.rdev-package}
\title{str.rdev: Streamline R Developer Case Study}
\description{
Download and clean data from one particular spreadsheet of the Enron Email Corpus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StreamlineDataScience/str.rdev}
  \item Report bugs at \url{https://github.com/StreamlineDataScience/str.rdev/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
