% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_data}
\alias{enron_download_data}
\title{Download the Enron XLS file}
\usage{
enron_download_data(destfile)
}
\arguments{
\item{destfile}{Destination for XLS file.}
}
\value{
A XLS file saved at the specified file path.
}
\description{
Download the Enron XLS file
}
\examples{
\dontrun{
enron_download_data("~/Desktop/enron_data.xls")
}
}
