% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_data}
\alias{enron_download_data}
\title{Download enron data}
\usage{
enron_download_data(
  source = c("github", "google_sheets"),
  link = NULL,
  destfile = "inst/extdata/enron_raw_data.xls"
)
}
\arguments{
\item{source}{Character. Indicates where to download enron data from.
Possible options are \code{"github"} and \code{"google_sheets"}.}

\item{link}{Character. Link to the \code{.xls} enron file.}

\item{destfile}{Character. Filepath where the \code{.xls} file should be stored.}
}
\description{
Download enron data
}
\details{
This function is used to donwload enron \code{.xls} data.
By default, i.e. when no arguments are passed to the function, it will
download data from GitHub and save it in \code{"inst/extdata/enron_raw_data.xls"}.
Default \code{link}s for different \code{source}s were defined in \code{R/constants.R}.
}
