% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_file}
\alias{enron_download_file}
\title{Download the Enron dataset.}
\usage{
enron_download_file()
}
\value{
The downloaded and initially cleaned Enron dataset from internet and ready to be processed further in R by other functions.
}
\description{
Download the Enron dataset.
}
\details{
This function downloads the Enron XLS file from github url and performs initial data cleaning.
}
\examples{
\dontrun{enron_download_file()}

}
