% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_file}
\alias{enron_download_file}
\title{Function for downloading and storing Enron data}
\usage{
enron_download_file(link = NULL)
}
\arguments{
\item{link}{A url to the Enron dataset to be downloaded.}
}
\value{
A raw XLS file is downloaded to the local temp file location via \code{\link{download.file}}.
}
\description{
Function for downloading and storing Enron data
}
\examples{
enron_download_file()
\dontrun{
enron_download_file()
}


}
