% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_download_data.R
\name{enron_download_file}
\alias{enron_download_file}
\title{Download Enron dataset}
\usage{
enron_download_file(filename = "enron")
}
\arguments{
\item{filename}{filename that is saved}
}
\value{
xls saved in /data-raw
}
\description{
Downloads enron dataset from the Streamline github repo and saves it locally
in /data-raw
}
