% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{enron_process_data}
\alias{enron_process_data}
\title{Process Enron spreadsheet data}
\usage{
enron_process_data(path)
}
\arguments{
\item{path}{The path to the spreadsheet.}
}
\value{
A tibble of clean data.
}
\description{
Process the spreadsheet of Enron data. This is currently quite fragile, and
will not function properly if new data is added to the spreadsheet, if rows
are combined, etc.
}
\examples{
enron_process_data(system.file("extdata/enron.xls", package = "str.rdev"))
}
