% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{enron_process_data}
\alias{enron_process_data}
\title{Processes Enron dataset to long format.}
\usage{
enron_process_data(file)
}
\arguments{
\item{file}{The file path of the \code{.xls} file for processing.}
}
\value{
Returns a \code{data.frame}.
}
\description{
Processes Enron dataset to long format.
}
\examples{
enron_process_data(enron_download_file())
}
