% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{enron_process_data}
\alias{enron_process_data}
\title{Process Enron data from XLS file}
\usage{
enron_process_data(file)
}
\arguments{
\item{file}{The file to process}
}
\value{
A tidy data frame
}
\description{
Process Enron data from XLS file
}
\examples{
\dontrun{
enron_process_data("~/Desktop/andrea_ring_000_1_1.pst.0.xls")
}
}
