% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{enron_process_data}
\alias{enron_process_data}
\title{Clean delivery/receipt exon dataset}
\usage{
enron_process_data(file = "exon")
}
\arguments{
\item{file}{name of raw .xls file to clean, is in /data-raw}
}
\value{
clean long df with
}
\description{
ccc
}
