% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{enron_process_data}
\alias{enron_process_data}
\title{Function to read and convert the raw Enron XLS file into a usable format}
\usage{
enron_process_data(
  location_names = c("ACADIAN", "BRIDGELINE", "COLUMBIA GULF", "DIGCO",
    "JEFFERSON ISLAND", "GULF SOUTH", "MAINLINE", "NGPL", "SONAT", "SEA ROBIN",
    "TEXAS GAS", "TRUNKLINE", "TRANSCO")
)
}
\arguments{
\item{location_names}{A vector of all location names that are present in the file. This will default to the original list of clients in the sample data.}
}
\value{
A processed dataframe is returned
}
\description{
Function to read and convert the raw Enron XLS file into a usable format
}
\examples{
enron_process_data()
\dontrun{
enron_process_data()
}


}
