% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enron_process_data.R
\name{enron_process_data}
\alias{enron_process_data}
\title{Process enron data}
\usage{
enron_process_data(file)
}
\arguments{
\item{file}{Character. Filepath of the \code{.xls} file to be processed.}
}
\description{
Process enron data
}
\details{
The structure of data stored in the \code{.xls} file is not ideal. Four
different elements were identified: locations, receipts, deliveries and dates.
The data processing was developed in a way that allows the user to easily
adapt code if the location of data changes. First, the set of different
location names is identified. This set of location names is later used in the
processing of receipts and deliveries, which is done via a helper function
that expects each data range with their corresponding dates. Finally,
all data is merged and arranged.
}
